cqfd: add cqfd start command
This adds 'cqfd start' command. Which is to start a container created
using the 'persistent' option.

Note that the command can have an optional container id. If it is not
provided first exited cqfd container is started.

To achieve this, a metadata is added to the container during its
creation. The metadata is named cqfd_entrypoint and it's value is the
path the entry-point script. While starting a container the metadata is
retrieved and the script is re-generated to use with the new container.